improve : multi thread flush

+++ tskv/src/tsm/index.rs
@@ -116,8 +116,7 @@ impl IndexMeta {
         if self.block_count == 0 {
             return (Timestamp::MIN, Timestamp::MIN);
         }
-        let first_blk_beg =
-            self.index_ref.field_id_offs()[self.index_idx].1 as usize + INDEX_META_SIZE;
+        let first_blk_beg = self.index_ref.field_id_offs()[self.index_idx].1 + INDEX_META_SIZE;
         let min_ts = decode_be_i64(&self.index_ref.data[first_blk_beg..first_blk_beg + 8]);
         let last_blk_beg = first_blk_beg + BLOCK_META_SIZE * (self.block_count as usize - 1);
         let max_ts = decode_be_i64(&self.index_ref.data[last_blk_beg + 8..last_blk_beg + 16]);
